Moved assert package into internal
